Create README.md
This repo is primarily to distribute Version 2 of Dealer a bridge hand generation program. The original version of Dealer was written by Hans van Staveren in the early 1990's. Several people contributed to the code over the years. Version 2 is a major upgrade to version 1.4 the version currently distributed by Debian.